requirements.txt
Cython==0.29.15
scikit-learn==0.22.1
pandas==1.0.1
tensorflow==2.1.0
pyjnius==1.3.0
docplex==2.11.176
tqdm
matplotlib